requirements-docker.txt
pandas
numpy
scikit-learn
statsmodels
xgboost
jinja2
mlflow
pydantic
typing
redis
fastapi
uvicorn
python-dotenv